+++ packages/lumx-react/src/components/date-picker/DatePicker.test.tsx
@@ -1,0 +1,97 @@
+import React, { ReactElement } from 'react';
+
+import moment from 'moment';
+
+import { mount, shallow } from 'enzyme';
+import 'jest-enzyme';
+
+import { ICommonSetup, Wrapper } from '@lumx/react/testing/utils';
+
+import { DatePicker, DatePickerProps } from './DatePicker';
+
+Date.now = jest.fn(() =>
+    new Date(
+        new Date(1487721600).toLocaleString('en-US', {
+            timeZone: 'America/Toronto',
+        }),
+    ).valueOf(),
+);
+
+/////////////////////////////
+
+/**
+ * Define the overriding properties waited by the `setup` function.
+ */
+type ISetupProps = Partial<DatePickerProps>;
+
+/**
+ * Defines what the `setup` function will return.
+ */
+interface ISetup extends ICommonSetup {
+    props: ISetupProps;
+}
+
+/////////////////////////////
+
+/**
+ * Mounts the component and returns common DOM elements / data needed in multiple tests further down.
+ *
+ * @param props  The props to use to override the default props of the component.
+ * @param     [shallowRendering=true] Indicates if we want to do a shallow or a full rendering.
+ * @return      An object with the props, the component wrapper and some shortcut to some element inside of the
+ *                       component.
+ */
+const setup = ({ ...propsOverrides }: ISetupProps = {}, shallowRendering: boolean = true): ISetup => {
+    const props: DatePickerProps = {
+        locale: 'fr',
+        onChange: jest.fn(),
+        value: moment(),
+        ...propsOverrides,
+    };
+
+    const renderer: (el: ReactElement) => Wrapper = shallowRendering ? shallow : mount;
+
+    const wrapper: Wrapper = renderer(<DatePicker {...props} />);
+
+    return {
+        props,
+        wrapper,
+    };
+};
+
+describe(`<${DatePicker.displayName}>`, (): void => {
+    // 1. Test render via snapshot (default states of component).
+    describe('Snapshots and structure', (): void => {
+        it('should render correctly', (): void => {
+            const { wrapper } = setup();
+            expect(wrapper).toMatchSnapshot();
+        });
+    });
+
+    /////////////////////////////
+
+    // 2. Test defaultProps value and important props custom values.
+    describe('Props', (): void => {
+        // Nothing to do here.
+    });
+
+    /////////////////////////////
+
+    // 3. Test events.
+    describe('Events', (): void => {
+        // Nothing to do here.
+    });
+    /////////////////////////////
+
+    // 4. Test conditions (i.e. things that display or not in the UI based on props).
+    describe('Conditions', (): void => {
+        // Nothing to do here.
+    });
+
+    /////////////////////////////
+
+    // 5. Test state.
+    describe('State', (): void => {
+        // Nothing to do here.
+    });
+});