+++ packages/lumx-react/src/components/date-picker/DatePickerField.tsx
@@ -1,10 +1,14 @@
-import React from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 
-import noop from 'lodash/noop';
+import { Placement, Popover, TextField, TextFieldProps, Theme } from '@lumx/react';
+import { useClickAway } from '@lumx/react/hooks/useClickAway';
+import { useFocusTrap } from '@lumx/react/hooks/useFocusTrap';
+import { onEscapePressed } from '@lumx/react/utils';
 
-import { Dropdown, Placement, TextField, Theme } from '@lumx/react';
+import { ENTER_KEY_CODE, SPACE_KEY_CODE } from '@lumx/react/constants';
+import { CLASSNAME, COMPONENT_NAME as COMPONENT_PREFIX, DatePicker, DatePickerProps } from './DatePicker';
 
-import { COMPONENT_NAME as COMPONENT_PREFIX, DatePicker, DatePickerProps } from './DatePicker';
+import { useFocus } from '@lumx/react/hooks/useFocus';
 
 /////////////////////////////
 
@@ -13,8 +17,14 @@ import { COMPONENT_NAME as COMPONENT_PREFIX, DatePicker, DatePickerProps } from 
  */
 
 type DatePickerFieldProps = DatePickerProps & {
+    /** Whether the text field is clearable. */
+    isClearable?: TextFieldProps['isClearable'];
+
     /** Input label. */
-    label: string;
+    label?: TextFieldProps['label'];
+
+    /** Text field placeholder message. */
+    placeholder?: string;
 
     /** Theme. */
     theme?: Theme;
@@ -33,6 +43,13 @@ type DatePickerFieldProps = DatePickerProps & {
  */
 const COMPONENT_NAME = `${COMPONENT_PREFIX}Field`;
 
+/**
+ * The default value of props.
+ */
+const DEFAULT_PROPS: Partial<DatePickerFieldProps> = {
+    isClearable: true,
+};
+
 /////////////////////////////
 
 /**
@@ -40,41 +57,114 @@ const COMPONENT_NAME = `${COMPONENT_PREFIX}Field`;
  *
  * @return The component.
  */
-const DatePickerField = ({ label, theme, value, ...props }: DatePickerFieldProps) => {
-    const anchorSimpleRef = React.useRef(null);
-    const [isSimpleOpen, setSimpleIsOpen] = React.useState(false);
+const DatePickerField = ({
+    isClearable = DEFAULT_PROPS.isClearable,
+    label,
+    placeholder,
+    theme,
+    value,
+    ...props
+}: DatePickerFieldProps) => {
+    const wrapperRef = useRef(null);
+    const popoverRef = useRef(null);
+    const anchorRef = useRef(null);
+
+    const [isOpen, setIsOpen] = useState(false);
 
     const toggleSimpleMenu = () => {
-        setSimpleIsOpen(!isSimpleOpen);
+        setIsOpen(!isOpen);
     };
 
     const closeSimpleMenu = () => {
-        setSimpleIsOpen(false);
+        setIsOpen(false);
+    };
+
+    const { computedPosition, isVisible } = Popover.useComputePosition(
+        Placement.BOTTOM_START!,
+        anchorRef,
+        popoverRef,
+        isOpen,
+        undefined,
+        false,
+        false,
+    );
+
+    const onEscapeHandler = isOpen && onEscapePressed(closeSimpleMenu);
+
+    useFocus(anchorRef.current, isOpen);
+    const handleKeyboardNav = (evt: React.KeyboardEvent<HTMLElement>): void => {
+        if ((evt.which === ENTER_KEY_CODE || evt.which === SPACE_KEY_CODE) && toggleSimpleMenu) {
+            toggleSimpleMenu();
+        }
+    };
+
+    useEffect(() => {
+        if (!onEscapeHandler || !wrapperRef.current) {
+            return undefined;
+        }
+        if (isOpen && wrapperRef.current) {
+            window.addEventListener('keydown', onEscapeHandler);
+        }
+        return (): void => {
+            window.removeEventListener('keydown', onEscapeHandler);
+        };
+    }, [isOpen, closeSimpleMenu]);
+
+    useClickAway(
+        wrapperRef,
+        () => {
+            if (!isOpen) {
+                return;
+            }
+
+            closeSimpleMenu();
+        },
+        [anchorRef],
+    );
+
+    // Handle focus trap.
+    const todayOrSelectedDateRef = useRef<HTMLButtonElement>(null);
+    useFocusTrap(todayOrSelectedDateRef.current && wrapperRef.current, todayOrSelectedDateRef.current);
+
+    const onChange = (textFieldValue: string) => {
+        if (!textFieldValue) {
+            props.onChange(undefined);
+        }
     };
 
     return (
         <>
             <TextField
-                textFieldRef={anchorSimpleRef}
+                isClearable={isClearable}
+                forceFocusStyle={isOpen}
+                textFieldRef={anchorRef}
                 label={label}
+                placeholder={placeholder}
                 value={value ? value.format('LL') : ''}
                 onClick={toggleSimpleMenu}
-                onChange={noop}
+                onChange={onChange}
+                onKeyPress={handleKeyboardNav}
                 theme={theme}
-                readOnly={true}
+                readOnly
             />
-
-            <Dropdown
-                showDropdown={isSimpleOpen}
-                closeOnClick={true}
-                closeOnEscape={true}
-                onClose={closeSimpleMenu}
-                placement={Placement.BOTTOM_START}
-                anchorRef={anchorSimpleRef}
-                fitToAnchorWidth={false}
-            >
-                <DatePicker value={value} {...props} />
-            </Dropdown>
+            {isOpen ? (
+                <Popover
+                    popoverRect={computedPosition}
+                    popoverRef={popoverRef}
+                    isVisible={isVisible}
+                    placement={Placement.BOTTOM_START}
+                >
+                    <div
+                        ref={wrapperRef}
+                        style={{
+                            maxHeight: computedPosition.maxHeight,
+                            minWidth: 0,
+                        }}
+                    >
+                        <DatePicker value={value} todayOrSelectedDateRef={todayOrSelectedDateRef} {...props} />
+                    </div>
+                </Popover>
+            ) : null}
         </>
     );
 };
@@ -82,4 +172,4 @@ DatePickerField.displayName = COMPONENT_NAME;
 
 /////////////////////////////
 
-export { COMPONENT_NAME, DatePickerField, DatePickerFieldProps };
+export { CLASSNAME, COMPONENT_NAME, DEFAULT_PROPS, DatePickerField, DatePickerFieldProps };